--- a/.enb/techs/html-from-bemtree.js
+++ b/.enb/techs/html-from-bemtree.js
@@ -6,9 +6,9 @@
  *
  * **Опции**
  *
- * * *String* **bemhtmlTarget** — Исходный BEMHTML-файл. По умолчанию — `?.bemhtml.js`.
- * * *String* **bemtreeTarget** — Исходный BEMJSON-файл. По умолчанию — `?.bemtree.js`.
- * * *String* **destTarget** — Результирующий HTML-файл. По умолчанию — `?.html`.
+ * * *String* **bemhtmlFile** — Исходный BEMHTML-файл. По умолчанию — `?.bemhtml.js`.
+ * * *String* **bemtreeFile** — Исходный BEMJSON-файл. По умолчанию — `?.bemtree.js`.
+ * * *String* **target** — Результирующий HTML-файл. По умолчанию — `?.html`.
  *
  * **Пример**
  *
@@ -16,36 +16,27 @@
  * nodeConfig.addTech(require('enb/techs/html-from-bemjson'));
  * ```
  */
-var vm = require('vm'),
-    vow = require('vow'),
-    vfs = require('enb/lib/fs/async-fs'),
+var vow = require('vow'),
     asyncRequire = require('enb/lib/fs/async-require'),
     dropRequireCache = require('enb/lib/fs/drop-require-cache');
 
 module.exports = require('enb/lib/build-flow').create()
     .name('html-from-bemtree')
-    .target('destTarget', '?.html')
-    .useSourceFilename('bemtreeTarget', '?.bemtree.js')
-    .useSourceFilename('bemhtmlTarget', '?.bemhtml.js')
+    .target('target', '?.html')
+    .useSourceFilename('bemtreeFile', '?.bemtree.js')
+    .useSourceFilename('bemhtmlFile', '?.bemhtml.js')
     .builder(function(bemtreeFilename, bemhtmlFilename) {
+        dropRequireCache(require, bemtreeFilename);
         dropRequireCache(require, bemhtmlFilename);
 
         return vow.all([
-                vfs.read(bemtreeFilename, 'utf-8'),
+                asyncRequire(bemtreeFilename),
                 asyncRequire(bemhtmlFilename)
             ])
             .spread(function(bemtree, bemhtml) {
-                var ctx = vm.createContext({
-                    Vow : vow,
-                    console : console,
-                    setTimeout : setTimeout
-                });
+                var BEMTREE = bemtree.BEMTREE,
+                    BEMHTML = bemhtml.BEMHTML;
 
-                vm.runInContext(bemtree, ctx);
-
-                return [ctx.BEMTREE, bemhtml.BEMHTML];
-            })
-            .spread(function(BEMTREE, BEMHTML) {
                 return BEMTREE.apply({})
                     .then(function(bemjson) {
                         return BEMHTML.apply(bemjson);