--- a/common.blocks/i-bem/__dom/_init/i-bem__dom_init_auto.deps.js
+++ b/common.blocks/i-bem/__dom/_init/i-bem__dom_init_auto.deps.js
@@ -1,3 +1,3 @@
 ({
-    shouldDeps : { elem : 'dom', mod : 'init' } // TODO: remove elem field after enb-bem/enb-bem-techs#158
+    shouldDeps : { mod : 'init' }
 })