--- a/package.json
+++ b/package.json
@@ -33,7 +33,7 @@
     "reqf": "^1.0.0",
     "bem-naming": "0.5.1",
     "enb": "^0.17.2",
-    "enb-bem-techs": "^1.0.4",
+    "enb-bem-techs": "^2.0.0",
     "enb-bem-docs": "0.8.0",
     "enb-bem-examples": "0.5.9",
     "enb-bem-specs": "0.5.6",