--- a/package.json
+++ b/package.json
@@ -36,7 +36,7 @@
     "enb-bem-techs": "^2.0.0",
     "enb-bem-docs": "0.8.1",
     "enb-bem-examples": "0.5.10",
-    "enb-bem-specs": "0.5.6",
+    "enb-bem-specs": "0.6.0",
     "enb-bem-tmpl-specs": "0.11.2",
     "enb-bemxjst": "1.3.4",
     "enb-bh": "0.5.0",