--- a/package.json
+++ b/package.json
@@ -34,7 +34,7 @@
     "bem-naming": "0.5.1",
     "enb": "^0.17.2",
     "enb-bem-techs": "^2.0.0",
-    "enb-bem-docs": "0.8.0",
+    "enb-bem-docs": "0.8.1",
     "enb-bem-examples": "0.5.9",
     "enb-bem-specs": "0.5.6",
     "enb-bem-tmpl-specs": "0.11.2",