--- a/package.json
+++ b/package.json
@@ -35,7 +35,7 @@
     "enb": "^0.17.2",
     "enb-bem-techs": "^2.0.0",
     "enb-bem-docs": "0.8.1",
-    "enb-bem-examples": "0.5.9",
+    "enb-bem-examples": "0.5.10",
     "enb-bem-specs": "0.5.6",
     "enb-bem-tmpl-specs": "0.11.2",
     "enb-bemxjst": "1.3.4",