--- a/package.json
+++ b/package.json
@@ -38,7 +38,7 @@
     "enb-bem-examples": "0.5.10",
     "enb-bem-specs": "0.6.0",
     "enb-bem-tmpl-specs": "0.13.2",
-    "enb-bemxjst": "1.3.4",
+    "enb-bemxjst": "2.0.0-rc",
     "enb-bh": "0.5.0",
     "enb-borschik": "^1.5.1",
     "enb-magic-platform": "0.5.0",