--- a/package.json
+++ b/package.json
@@ -40,7 +40,7 @@
     "enb-bem-tmpl-specs": "0.13.2",
     "enb-bemxjst": "2.0.0-rc",
     "enb-bh": "0.5.0",
-    "enb-borschik": "^1.5.1",
+    "enb-borschik": "2.0.0-beta1",
     "enb-magic-platform": "0.5.0",
     "enb-modules": "0.2.0",
     "borschik-tech-cleancss": "2.0.1"