--- a/package.json
+++ b/package.json
@@ -39,7 +39,7 @@
     "enb-bem-specs": "0.6.0",
     "enb-bem-tmpl-specs": "0.13.2",
     "enb-bemxjst": "2.0.0-rc",
-    "enb-bh": "0.5.0",
+    "enb-bh": "1.0.0-rc2",
     "enb-borschik": "2.0.0-beta1",
     "enb-magic-platform": "0.5.0",
     "enb-modules": "0.2.0",