Resolved #1124 

How the speed has changed is shown in table.

 | Before | After
--------- | -------------- | ------
dist | 9953ms | **8768ms**
specs | 29198ms | 29079ms
tmpl-specs | **2083ms** | 2468ms
docs | 3719ms | 3803ms
examples | 18134ms | **8833ms**

**Conclusion**

* :rocket: accelerate the assembly of examples (`enb-bemxjst` is faster)
* :heavy_plus_sign: accelerate the assembly of dist (one BH-file less)
* :interrobang: slightly slowed down assembly of `tmpl-specs` (package itself)